removi rota e pasta new-project/peguei a logica newProject e coloquei no btn
