+++ src/routes/routes.jsx
@@ -8,7 +8,6 @@ import UserProfile from "../views/UserProfile";
 import NicknameInput from "../views/NicknameInput";
 import ProjectPage from "../views/ProjectPage";
 import ForgotPassword from "../views/ForgotPassword"
-import NewProject from "../views/NewProject";
 import Gallery from "../views/Gallery";
 import UserConfig from "../views/UserConfig";
 
@@ -26,7 +25,6 @@ export default function Rotas() {
                 <Route path="/profile/nickname" element={<NicknameInput />} />
                 <Route path="/project/:ProjectID" element={<ProjectPage />} />
                 <Route path="/forgotpassword" element={<ForgotPassword />} />
-                <Route path="/new-project" element={<NewProject />} />
                 <Route path="/gallery" element={<Gallery />} />
                 <Route path="/Userconfig" element={<UserConfig />} />
             </Routes>
